tests: run stunnel for HTTPS and FTPS on dynamic ports
stunnel is an external tool and it has no specific option to export the
actually used port number. Runtests tries to figure it out by parsing
the stunnel log. This is fragile. As a backup, it runs "netstat -ptln"
in an attempt to see what stunnel listens to. That's not going to be
widely portable.

Closes #5267